+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/KubernetesClient.java
@@ -75,7 +75,6 @@ import io.fabric8.kubernetes.client.dsl.NamespaceListVisitFromServerGetDeleteRec
 import io.fabric8.kubernetes.client.dsl.NamespaceableResource;
 import io.fabric8.kubernetes.client.dsl.NetworkAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.NonNamespaceOperation;
-import io.fabric8.kubernetes.client.dsl.ParameterNamespaceListVisitFromServerGetDeleteRecreateWaitApplicable;
 import io.fabric8.kubernetes.client.dsl.PodResource;
 import io.fabric8.kubernetes.client.dsl.PolicyAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.RbacAPIGroupDSL;
@@ -291,7 +290,7 @@ public interface KubernetesClient extends Client {
    * @param is the input stream containing JSON/YAML content
    * @return an operation instance to work on the list of Kubernetes Resource objects
    */
-  ParameterNamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata> load(InputStream is);
+  NamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata> load(InputStream is);
 
   /**
    * Load a Kubernetes list object
@@ -299,7 +298,7 @@ public interface KubernetesClient extends Client {
    * @param s kubernetes list as string
    * @return an operation instance to work on the deserialized KubernetesList objects
    */
-  ParameterNamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata> resourceList(String s);
+  NamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata> resourceList(String s);
 
   /**
    * KubernetesResourceList operations